pyproject.toml
[project]
name = "xero-expense-audit"
version = "0.1.0"
description = "Xero expense validation — deterministic rules + Claude Haiku"
requires-python = ">=3.12"
dependencies = [
  "xero-python>=6.0.0",
  "anthropic>=0.21.0",
  "click>=8.1.0",
  "python-dotenv>=1.0.0",
  "httpx>=0.27.0",
  "pyyaml>=6.0.1",
  "rich>=13.7.0",
  "tabulate>=0.9.0",
  "pypdf>=4.0.0",
  "questionary>=2.1.1",
]

[dependency-groups]
dev = ["pytest>=8.0.0", "pytest-httpx>=0.30.0"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
